// src/modules/companies/dto/create-company.dto.ts
import { $Enums, company } from '@prisma/client';

export class CreateCompanyDto implements company {
	max_reach_level: $Enums.reach_level;
	name: string;
	id: string;
	slug: string;
	cnpj: string | null;
	cover_img_url: string | null;
	description: string | null;
	active: boolean;
	created_at: Date;
	updated_at: Date;
}